examples/anthropic-basic/pyproject.toml
[project]
name = "example-anthropic-basic"
version = "0.0.0"
description = "Basic Anthropic agent example"
requires-python = ">=3.10"
dependencies = [
    "reminix-runtime",
    "reminix-anthropic",
    "anthropic>=0.75.0",
    "python-dotenv>=1.0.0",
]

[tool.uv.sources]
reminix-runtime = { workspace = true }
reminix-anthropic = { workspace = true }
